0.8.18.14: 	Merge x86-64-again branch onto HEAD.
	Many, many, many 64-bit cleanups in code/, runtime/, compiler/,
	compiler/generic/

	New SAP-REF-WORD and friends.

	Various fixes to the x86-64 backends (and addition of assembly/
	and runtime/ files necessary).  Implementation of Unicode-related
	stuff by CSR.  Signed modular arithmetic has not yet been
	implemented.

	A number of tests fail:
	... alien.impure.lisp: enum <-> integer array conversion
	... exhaust.impure.lisp: "deferred gubbins"
	... float.pure.lisp: float infinities
	... foreign.test.sh: "deferred gubbins"

	It's possible that this merge will cause alpha32 to break in an
	interesting way, probably related to undefined-alien.  Needs
	debugging.  Other architectures have been tested, but of course
	it's possible that something has gone wrong.

	Though I (CSR) am merging this, the vast majority of the work was
	done by Juho Snellman (building on Dan Barlow's initial work to
	get it into executing lisp code in cold-init), with guest appearances
	by Cheuksan Edward Wang and Vincent Arkesteijn.

+++ src/runtime/gencgc-alloc-region.h
@@ -9,14 +9,14 @@ struct alloc_region {
     void  *end_addr; /* pointer to the byte after the last usable byte */
 
     /* These are needed when closing the region. */
-    int  first_page;
-    int  last_page;
+    long  first_page;
+    long  last_page;
     void  *start_addr;
 };
 
 extern struct alloc_region  boxed_region;
 extern struct alloc_region  unboxed_region;
-extern int from_space, new_space;
+extern long from_space, new_space;
 extern struct weak_pointer *weak_pointers;
 
 extern void *current_region_free_pointer;
